libselinux: Support Android/M and official v30 sepolicy format
Some Android M devices use a slightly different format for the sepolicy
file.  Compatibility for this is not present in SELinuxProject/selinux
master.  Google does provide a patch for the kernel itself to be able to
handle both formats - https://android-review.googlesource.com/#/c/179568/ -
but not for libsepol. This patch is a port of the kernel patch to
libsepol.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>